Add files via upload
PROBLEM STATEMENT
Develop a portal indicating the summary of Fitness Certificate categories.
Number of NO's in the corresponding FC must be displayed. Duration of FC
submission is bimonthly. number of 3 consecutive NO's in the subcategory of each FC
needs to be displayed and the corresponding list needs to be popped up.
I have just started developing the project workflow to gather and address the necessary requirements.